services/agent/requirements.txt
kafka-python==2.0.2
pandas==2.2.2
pyarrow==15.0.2
